Inline dom::eventdispatcher into dom::event
